+++ encore_api_cli/commands/analysis.py
@@ -1,6 +1,9 @@
+import json
+
 import click
 
 from encore_api_cli.options import common_options
+from encore_api_cli.output import write_json_data, write_message
 from encore_api_cli.state import pass_state
 from encore_api_cli.utils import get_client
 
@@ -12,8 +15,23 @@ def cli() -> None:  # noqa: D103
 
 @cli.group()
 def analysis():
-    """Manege analyses."""
-    pass
+    """Show analysis results."""
+
+
+@analysis.command()
+@click.argument("analysis_id", type=int)
+@common_options
+@pass_state
+def show(state, analysis_id):
+    """Show analysis result."""
+    c = get_client(state)
+    response = c.get_info("analyses", analysis_id)
+    status = response.get("exec_status", "FAILURE")
+    if status == "SUCCESS":
+        result = response.get("result")
+        write_json_data(json.loads(result), sort_keys=False)
+    else:
+        write_message("Status is not SUCCESS.")
 
 
 @analysis.command()
@@ -22,15 +40,4 @@ def analysis():
 def list(state):
     """Show analysis list."""
     c = get_client(state)
-    c.show_list("analyses")
-
-
-@analysis.command()
-@click.argument("analysis_id", type=int)
-@common_options
-@pass_state
-def show(state, analysis_id):
-    """Display analysis result in JSON format."""
-    c = get_client(state)
-    result = c.get_analysis(analysis_id)
-    click.echo(result)
+    write_json_data(c.get_info("analyses"), sort_keys=False)