+++ encore_api_cli/commands/image.py
@@ -1,6 +1,7 @@
 import click
 
 from encore_api_cli.options import common_options
+from encore_api_cli.output import write_json_data
 from encore_api_cli.state import pass_state
 from encore_api_cli.utils import get_client
 
@@ -12,14 +13,24 @@ def cli() -> None:  # noqa: D103
 
 @cli.group()
 def image():
-    """Manege images."""
+    """Show the information of the uploaded images."""
     pass
+
+
+@image.command()
+@click.argument("image_id", type=int)
+@common_options
+@pass_state
+def show(state, image_id):
+    """Show image information."""
+    c = get_client(state)
+    write_json_data(c.get_info("images", image_id), sort_keys=False)
 
 
 @image.command()
 @common_options
 @pass_state
 def list(state):
-    """Show image list."""
+    """Show a list of information for all images."""
     c = get_client(state)
-    c.show_list("images")
+    write_json_data(c.get_info("images"), sort_keys=False)