+++ encore_api_cli/commands/movie.py
@@ -1,6 +1,7 @@
 import click
 
 from encore_api_cli.options import common_options
+from encore_api_cli.output import write_json_data
 from encore_api_cli.state import pass_state
 from encore_api_cli.utils import get_client
 
@@ -12,14 +13,24 @@ def cli() -> None:  # noqa: D103
 
 @cli.group()
 def movie():
-    """Manege movies."""
+    """Show the information of the uploaded movies."""
     pass
+
+
+@movie.command()
+@click.argument("movie_id", type=int)
+@common_options
+@pass_state
+def show(state, movie_id):
+    """Show movie information."""
+    c = get_client(state)
+    write_json_data(c.get_info("movies", movie_id), sort_keys=False)
 
 
 @movie.command()
 @common_options
 @pass_state
 def list(state):
-    """Show movie list."""
+    """Show a list of information for all movies."""
     c = get_client(state)
-    c.show_list("movies")
+    write_json_data(c.get_info("movies"), sort_keys=False)