+++ encore_api_cli/commands/upload.py
@@ -4,7 +4,7 @@ from encore_api_cli.exceptions import InvalidFileType, RequestsError
 from encore_api_cli.options import common_options
 from encore_api_cli.output import write_success
 from encore_api_cli.state import State, pass_state
-from encore_api_cli.utils import get_client
+from encore_api_cli.utils import color_id, color_path, get_client
 
 
 @click.group()
@@ -27,10 +27,6 @@ def upload(state: State, path: str) -> None:
     except RequestsError as e:
         raise click.ClickException(str(e))
 
-    write_success(
-        "Uploaded {path} to the cloud storage. ({media_type}_id: {media_id})".format(
-            path=click.style(path, fg="blue"),
-            media_type=media_type,
-            media_id={click.style(str(media_id), fg="cyan")},
-        )
-    )
+    cpath = color_path(path)
+    cid = color_id(media_id)
+    write_success(f"Uploaded {cpath} to the cloud storage. ({media_type}_id: {cid})")