+++ encore_api_cli/utils.py
@@ -33,3 +33,13 @@ def get_settings(profile: str, use_env: bool = True) -> Settings:
     except SettingsValueError as e:
         raise click.ClickException(str(e))
     return settings
+
+
+def color_id(number: int) -> str:
+    """Set color to ID."""
+    return click.style(str(number), fg="cyan")
+
+
+def color_path(path: str) -> str:
+    """Set color to path."""
+    return click.style(str(path), fg="blue")